Updated README
Updated the README file to reflect changes to sync_radar_data_globus.py.
Users are no longer required to get a client ID or set up a globus sdk project. The script also now grabs the personal UUID of the user, so they don't need to manually insert that either.
Python 3 is required to run this script.